ast_results/mtotschnig_MyExpenses/myExpenses/src/main/java/org/totschnig/myexpenses/activity/MyExpenses.java
// isComment
package org.totschnig.myexpenses.activity;

import android.app.ProgressDialog;
import android.arch.lifecycle.ViewModelProviders;
import android.content.ContentUris;
import android.content.Context;
import android.content.Intent;
import android.content.res.Configuration;
import android.content.res.Resources;
import android.database.Cursor;
import android.net.Uri;
import android.os.Build;
import android.os.Bundle;
import android.support.annotation.NonNull;
import android.support.annotation.Nullable;
import android.support.design.widget.NavigationView;
import android.support.design.widget.Snackbar;
import android.support.v4.app.Fragment;
import android.support.v4.app.FragmentManager;
import android.support.v4.app.LoaderManager;
import android.support.v4.content.CursorLoader;
import android.support.v4.content.Loader;
import android.support.v4.view.GravityCompat;
import android.support.v4.view.ViewPager;
import android.support.v4.view.ViewPager.OnPageChangeListener;
import android.support.v4.widget.DrawerLayout;
import android.support.v7.app.ActionBarDrawerToggle;
import android.support.v7.widget.Toolbar;
import android.text.ClipboardManager;
import android.util.TypedValue;
import android.view.ContextMenu;
import android.view.Menu;
import android.view.MenuInflater;
import android.view.MenuItem;
import android.view.View;
import android.view.ViewGroup;
import android.view.ViewTreeObserver;
import android.view.Window;
import android.view.WindowManager;
import android.widget.AdapterView;
import org.apache.commons.lang3.ArrayUtils;
import org.totschnig.myexpenses.MyApplication;
import org.totschnig.myexpenses.R;
import org.totschnig.myexpenses.adapter.MyGroupedAdapter;
import org.totschnig.myexpenses.dialog.BalanceDialogFragment;
import org.totschnig.myexpenses.dialog.ConfirmationDialogFragment;
import org.totschnig.myexpenses.dialog.ConfirmationDialogFragment.ConfirmationDialogListener;
import org.totschnig.myexpenses.dialog.ExportDialogFragment;
import org.totschnig.myexpenses.dialog.MessageDialogFragment;
import org.totschnig.myexpenses.dialog.ProgressDialogFragment;
import org.totschnig.myexpenses.dialog.RemindRateDialogFragment;
import org.totschnig.myexpenses.dialog.SortUtilityDialogFragment;
import org.totschnig.myexpenses.dialog.TransactionDetailFragment;
import org.totschnig.myexpenses.fragment.ContextualActionBarFragment;
import org.totschnig.myexpenses.fragment.TransactionList;
import org.totschnig.myexpenses.model.Account;
import org.totschnig.myexpenses.model.AccountGrouping;
import org.totschnig.myexpenses.model.AccountType;
import org.totschnig.myexpenses.model.ContribFeature;
import org.totschnig.myexpenses.model.CurrencyUnit;
import org.totschnig.myexpenses.model.Grouping;
import org.totschnig.myexpenses.model.Money;
import org.totschnig.myexpenses.model.Sort;
import org.totschnig.myexpenses.model.SortDirection;
import org.totschnig.myexpenses.model.Template;
import org.totschnig.myexpenses.model.Transaction;
import org.totschnig.myexpenses.preference.PrefKey;
import org.totschnig.myexpenses.provider.DatabaseConstants;
import org.totschnig.myexpenses.provider.TransactionProvider;
import org.totschnig.myexpenses.provider.filter.CommentCriteria;
import org.totschnig.myexpenses.provider.filter.Criteria;
import org.totschnig.myexpenses.task.TaskExecutionFragment;
import org.totschnig.myexpenses.ui.CursorFragmentPagerAdapter;
import org.totschnig.myexpenses.ui.FragmentPagerAdapter;
import org.totschnig.myexpenses.util.AppDirHelper;
import org.totschnig.myexpenses.util.ColorUtils;
import org.totschnig.myexpenses.util.CurrencyFormatter;
import org.totschnig.myexpenses.util.DistribHelper;
import org.totschnig.myexpenses.util.Result;
import org.totschnig.myexpenses.util.ShareUtils;
import org.totschnig.myexpenses.util.TextUtils;
import org.totschnig.myexpenses.util.UiUtils;
import org.totschnig.myexpenses.util.Utils;
import org.totschnig.myexpenses.util.ads.AdHandler;
import org.totschnig.myexpenses.viewmodel.RoadmapViewModel;
import java.io.Serializable;
import java.util.AbstractMap;
import java.util.ArrayList;
import java.util.Collections;
import java.util.Locale;
import javax.inject.Inject;
import butterknife.BindView;
import butterknife.ButterKnife;
import eltos.simpledialogfragment.SimpleDialog;
import eltos.simpledialogfragment.input.SimpleInputDialog;
import eltos.simpledialogfragment.list.MenuDialog;
import se.emilsjolander.stickylistheaders.ExpandableStickyListHeadersListView;
import se.emilsjolander.stickylistheaders.StickyListHeadersListView;
import timber.log.Timber;
import static eltos.simpledialogfragment.list.CustomListDialog.SELECTED_SINGLE_ID;
import static org.totschnig.myexpenses.contract.TransactionsContract.Transactions.OPERATION_TYPE;
import static org.totschnig.myexpenses.contract.TransactionsContract.Transactions.TYPE_TRANSACTION;
import static org.totschnig.myexpenses.provider.DatabaseConstants.KEY_ACCOUNTID;
import static org.totschnig.myexpenses.provider.DatabaseConstants.KEY_CLEARED_TOTAL;
import static org.totschnig.myexpenses.provider.DatabaseConstants.KEY_COLOR;
import static org.totschnig.myexpenses.provider.DatabaseConstants.KEY_CURRENCY;
import static org.totschnig.myexpenses.provider.DatabaseConstants.KEY_CURRENT_BALANCE;
import static org.totschnig.myexpenses.provider.DatabaseConstants.KEY_GROUPING;
import static org.totschnig.myexpenses.provider.DatabaseConstants.KEY_HAS_CLEARED;
import static org.totschnig.myexpenses.provider.DatabaseConstants.KEY_HAS_EXPORTED;
import static org.totschnig.myexpenses.provider.DatabaseConstants.KEY_LABEL;
import static org.totschnig.myexpenses.provider.DatabaseConstants.KEY_RECONCILED_TOTAL;
import static org.totschnig.myexpenses.provider.DatabaseConstants.KEY_ROWID;
import static org.totschnig.myexpenses.provider.DatabaseConstants.KEY_SECOND_GROUP;
import static org.totschnig.myexpenses.provider.DatabaseConstants.KEY_SORT_KEY;
import static org.totschnig.myexpenses.provider.DatabaseConstants.KEY_TRANSACTIONID;
import static org.totschnig.myexpenses.provider.DatabaseConstants.KEY_YEAR;
import static org.totschnig.myexpenses.task.TaskExecutionFragment.KEY_LONG_IDS;
import static org.totschnig.myexpenses.task.TaskExecutionFragment.TASK_EXPORT;
import static org.totschnig.myexpenses.task.TaskExecutionFragment.TASK_PRINT;

/**
 * isComment
 */
public class isClassOrIsInterface extends LaunchActivity implements OnPageChangeListener, LoaderManager.LoaderCallbacks<Cursor>, ConfirmationDialogFragment.ConfirmationDialogCheckedListener, ConfirmationDialogListener, ContribIFace, SimpleDialog.OnDialogResultListener, SortUtilityDialogFragment.OnConfirmListener {

    public static final long isVariable = isStringConstant;

    public static final int isVariable = -isIntegerConstant;

    public static final String isVariable = "isStringConstant";

    private static final String isVariable = "isStringConstant";

    private static final String isVariable = "isStringConstant";

    private LoaderManager isVariable;

    private int isVariable = -isIntegerConstant;

    private Cursor isVariable;

    private MyViewPagerAdapter isVariable;

    private MyGroupedAdapter isVariable;

    private long isVariable = isIntegerConstant;

    private String isVariable;

    private int isVariable = isIntegerConstant;

    private AdHandler isVariable;

    private Toolbar isVariable;

    private String isVariable;

    public enum HelpVariant {

        crStatus
    }

    private void isMethod() {
        Account isVariable = isNameExpr.isMethod(isNameExpr);
        isMethod(isNameExpr == null || isNameExpr.isMethod().isMethod(isNameExpr.isFieldAccessExpr) ? null : isNameExpr.isFieldAccessExpr);
    }

    /**
     * isComment
     */
    private long isVariable = isIntegerConstant;

    @BindView(isNameExpr.isFieldAccessExpr.isFieldAccessExpr)
    ExpandableStickyListHeadersListView isVariable;

    @Nullable
    @BindView(isNameExpr.isFieldAccessExpr.isFieldAccessExpr)
    DrawerLayout isVariable;

    @BindView(isNameExpr.isFieldAccessExpr.isFieldAccessExpr)
    ViewPager isVariable;

    @BindView(isNameExpr.isFieldAccessExpr.isFieldAccessExpr)
    NavigationView isVariable;

    private ActionBarDrawerToggle isVariable;

    private int isVariable, isVariable, isVariable, isVariable;

    boolean isVariable = true;

    private long isVariable = isIntegerConstant;

    private String isVariable = null;

    @Inject
    CurrencyFormatter isVariable;

    private RoadmapViewModel isVariable;

    @Override
    protected void isMethod() {
        isNameExpr.isMethod().isMethod().isMethod(this);
    }

    @Override
    public void isMethod(Bundle isParameter) {
        isMethod(isMethod());
        super.isMethod(isNameExpr);
        isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        final ViewGroup isVariable = isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        isNameExpr = isNameExpr.isMethod(isNameExpr);
        isNameExpr.isMethod().isMethod(new ViewTreeObserver.OnGlobalLayoutListener() {

            @Override
            public void isMethod() {
                isNameExpr.isMethod().isMethod(this);
                isNameExpr.isMethod();
            }
        });
        isNameExpr.isMethod();
        isNameExpr.isMethod(this);
        isNameExpr = isMethod(true);
        isNameExpr.isMethod(isParameter -> isMethod());
        if (isNameExpr != null) {
            isNameExpr = new ActionBarDrawerToggle(this, isNameExpr, isNameExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr) {

                /**
                 * isComment
                 */
                public void isMethod(View isParameter) {
                    super.isMethod(isNameExpr);
                    TransactionList isVariable = isMethod();
                    if (isNameExpr != null)
                        isNameExpr.isMethod();
                }

                /**
                 * isComment
                 */
                public void isMethod(View isParameter) {
                    super.isMethod(isNameExpr);
                    TransactionList isVariable = isMethod();
                    if (isNameExpr != null)
                        isNameExpr.isMethod();
                }

                @Override
                public void isMethod(View isParameter, float isParameter) {
                    // isComment
                    super.isMethod(isNameExpr, isIntegerConstant);
                }
            };
            // isComment
            isNameExpr.isMethod(isNameExpr);
        }
        isNameExpr = new MyGroupedAdapter(this, null, isNameExpr, isMethod(), isNameExpr);
        isNameExpr.isMethod(isParameter -> isMethod(isNameExpr.isMethod(), null));
        View isVariable = isNameExpr.isMethod(isIntegerConstant);
        if (isNameExpr != null) {
            isNameExpr.isMethod(true);
        }
        isNameExpr.isMethod(isNameExpr);
        isNameExpr.isMethod(true);
        isNameExpr.isMethod(new StickyListHeadersListView.OnHeaderClickListener() {

            @Override
            public void isMethod(StickyListHeadersListView isParameter, View isParameter, int isParameter, long isParameter, boolean isParameter) {
                if (isNameExpr.isMethod(isNameExpr)) {
                    isNameExpr.isMethod(isNameExpr);
                } else {
                    isNameExpr.isMethod(isNameExpr);
                }
            }

            @Override
            public boolean isMethod(StickyListHeadersListView isParameter, View isParameter, int isParameter, long isParameter, boolean isParameter) {
                return true;
            }
        });
        isNameExpr.isMethod((isParameter, isParameter, isParameter, isParameter) -> {
            if (isNameExpr != isNameExpr) {
                isMethod(isNameExpr);
                isMethod();
            }
        });
        isMethod(isNameExpr);
        isMethod(isNameExpr.isMethod(this, "isStringConstant", isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr));
        if (isNameExpr != null) {
            isNameExpr = isNameExpr.isMethod("isStringConstant");
            isNameExpr = isNameExpr.isMethod(isNameExpr, isStringConstant);
        } else {
            Bundle isVariable = isMethod().isMethod();
            if (isNameExpr != null) {
                isNameExpr = isNameExpr.isMethod(isNameExpr, isNameExpr, isIntegerConstant);
                isNameExpr = isNameExpr.isMethod(isNameExpr, isIntegerConstant);
                // isComment
                if (isNameExpr != isIntegerConstant) {
                    FragmentManager isVariable = isMethod();
                    if (isNameExpr.isMethod(TransactionDetailFragment.class.isMethod()) == null) {
                        isNameExpr.isMethod(isNameExpr).isMethod(isNameExpr, TransactionDetailFragment.class.isMethod());
                        isMethod().isMethod(isNameExpr);
                    }
                }
            }
        }
        if (isNameExpr == isIntegerConstant) {
            isNameExpr = isNameExpr.isFieldAccessExpr.isMethod(isStringConstant);
        }
        isNameExpr = isNameExpr.isMethod(this).isMethod(RoadmapViewModel.class);
        isMethod();
        if (isNameExpr == null) {
            isMethod();
        }
    }

    private void isMethod() {
        isMethod();
        Resources.Theme isVariable = isMethod();
        TypedValue isVariable = new TypedValue();
        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr, true);
        isNameExpr = new MyViewPagerAdapter(this, isMethod(), null);
        isNameExpr.isMethod(this.isFieldAccessExpr);
        isNameExpr.isMethod(this);
        isNameExpr.isMethod((int) isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isIntegerConstant, isMethod().isMethod()));
        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr);
        isNameExpr = isMethod();
        isNameExpr.isMethod(isNameExpr, null, this);
    }

    private void isMethod() {
        final String isVariable = "isStringConstant" + isNameExpr.isFieldAccessExpr;
        if (isNameExpr.isMethod(this) > isIntegerConstant && !isMethod().isMethod(isNameExpr, true)) {
            isNameExpr.isMethod().isMethod(this, isParameter -> {
                boolean isVariable = isNameExpr == null;
                if (isNameExpr || isNameExpr.isMethod() < isNameExpr.isFieldAccessExpr) {
                    Bundle isVariable = new Bundle();
                    isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isNameExpr ? isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr) : isNameExpr.isMethod(isNameExpr.this, "isStringConstant", isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr));
                    isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                    isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isNameExpr);
                    isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                    isNameExpr.isMethod(isNameExpr).isMethod(isMethod(), "isStringConstant");
                }
            });
            isNameExpr.isMethod();
        }
    }

    private void isMethod(int isParameter) {
        if (isNameExpr.isMethod() == isNameExpr)
            isMethod(isNameExpr);
        else
            isNameExpr.isMethod(isNameExpr, true);
    }

    private AccountGrouping isMethod() {
        try {
            return isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isMethod("isStringConstant"));
        } catch (IllegalArgumentException isParameter) {
            return isNameExpr.isFieldAccessExpr;
        }
    }

    private Sort isMethod() {
        try {
            return isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isMethod("isStringConstant"));
        } catch (IllegalArgumentException isParameter) {
            return isNameExpr.isFieldAccessExpr;
        }
    }

    @Override
    public boolean isMethod(Menu isParameter) {
        MenuInflater isVariable = isMethod();
        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr);
        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr);
        super.isMethod(isNameExpr);
        return true;
    }

    @Override
    public void isMethod(ContextMenu isParameter, View isParameter, ContextMenu.ContextMenuInfo isParameter) {
        if (((AdapterView.AdapterContextMenuInfo) isNameExpr).isFieldAccessExpr > isIntegerConstant) {
            isNameExpr.isMethod(isIntegerConstant, isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isIntegerConstant, isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
            if (isNameExpr.isMethod() > isIntegerConstant) {
                isNameExpr.isMethod(isIntegerConstant, isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isIntegerConstant, isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
            }
        }
    }

    @Override
    public boolean isMethod(MenuItem isParameter) {
        isMethod(isNameExpr.isMethod(), ((AdapterView.AdapterContextMenuInfo) isNameExpr.isMethod()).isFieldAccessExpr);
        return true;
    }

    /*isComment*/
    @Override
    protected void isMethod(int isParameter, int isParameter, Intent isParameter) {
        super.isMethod(isNameExpr, isNameExpr, isNameExpr);
        if (isNameExpr == isNameExpr && isNameExpr == isNameExpr) {
            long isVariable;
            isNameExpr = isNameExpr.isMethod(isNameExpr, isIntegerConstant);
            if (!isNameExpr.isMethod()) {
                isNameExpr = isNameExpr.isFieldAccessExpr.isMethod(isNameExpr);
                if (isNameExpr != -isIntegerConstant && isNameExpr >= isNameExpr) {
                    RemindRateDialogFragment isVariable = new RemindRateDialogFragment();
                    isNameExpr.isMethod(true);
                    isNameExpr.isMethod(isMethod(), "isStringConstant");
                    return;
                }
            }
            isNameExpr.isMethod();
        }
        if (isNameExpr == isNameExpr && isNameExpr == isNameExpr) {
            isNameExpr = isNameExpr.isMethod(isNameExpr, isIntegerConstant);
        }
    }

    public void isMethod(Integer isParameter, Criteria isParameter) {
        TransactionList isVariable = isMethod();
        if (isNameExpr != null) {
            isNameExpr.isMethod(isNameExpr, isNameExpr);
        }
    }

    /**
     * isComment
     */
    private void isMethod() {
        Intent isVariable = new Intent(this, ExpenseEdit.class);
        isNameExpr.isMethod(isNameExpr, isNameExpr);
        // isComment
        if (isNameExpr < isIntegerConstant) {
            isNameExpr.isMethod(isNameExpr, isNameExpr);
            isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, true);
        } else {
            // isComment
            isNameExpr.isMethod(isNameExpr, isNameExpr);
        }
        isMethod(isNameExpr, isNameExpr);
    }

    @Override
    protected void isMethod(String isParameter) {
        isMethod();
        super.isMethod(isNameExpr);
    }

    /**
     * isComment
     */
    public boolean isMethod(int isParameter, Object isParameter) {
        if (super.isMethod(isNameExpr, isNameExpr)) {
            return true;
        }
        Intent isVariable;
        TransactionList isVariable;
        switch(isNameExpr) {
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isMethod(isNameExpr.isFieldAccessExpr, null);
                return true;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isNameExpr = isMethod();
                if (isNameExpr != null && isNameExpr.isMethod()) {
                    isMethod(isNameExpr.isFieldAccessExpr, null);
                } else {
                    isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                }
                return true;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isNameExpr = isMethod();
                if (isNameExpr != null && isNameExpr.isMethod()) {
                    isMethod(isNameExpr.isFieldAccessExpr, null);
                } else {
                    isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                }
                return true;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                if (isNameExpr == isIntegerConstant) {
                    isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr);
                } else {
                    isMethod();
                }
                return true;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isNameExpr = isMethod();
                if (isNameExpr != null && isMethod()) {
                    isNameExpr.isMethod(isNameExpr);
                    CurrencyUnit isVariable = isNameExpr.isMethod(isNameExpr);
                    Bundle isVariable = new Bundle();
                    isNameExpr.isMethod(isNameExpr, isNameExpr.isMethod(isNameExpr));
                    isNameExpr.isMethod(isNameExpr, isNameExpr.isMethod(isNameExpr));
                    isNameExpr.isMethod(isNameExpr, isNameExpr.isMethod(new Money(isNameExpr, isNameExpr.isMethod(isNameExpr.isMethod(isNameExpr)))));
                    isNameExpr.isMethod(isNameExpr, isNameExpr.isMethod(new Money(isNameExpr, isNameExpr.isMethod(isNameExpr.isMethod(isNameExpr)))));
                    isNameExpr.isMethod(isNameExpr).isMethod(isMethod(), "isStringConstant");
                } else {
                    isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                }
                return true;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isNameExpr = isMethod();
                if (isNameExpr != null && isNameExpr.isMethod()) {
                    Result isVariable = isNameExpr.isMethod(this);
                    if (isNameExpr.isMethod()) {
                        isNameExpr.isMethod(isNameExpr, isNameExpr.isMethod()).isMethod(this.isMethod(), "isStringConstant");
                    } else {
                        isMethod(isNameExpr.isMethod(this), isNameExpr.isFieldAccessExpr);
                    }
                } else {
                    isMethod();
                }
                return true;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isNameExpr.isFieldAccessExpr.isMethod(-isIntegerConstant);
                return true;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isNameExpr.isFieldAccessExpr.isMethod(isNameExpr + isNameExpr);
                return true;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isNameExpr = new Intent(this, Help.class);
                isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, "isStringConstant");
                // isComment
                isMethod(isNameExpr);
                return true;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isNameExpr = new Intent(this, ManageTemplates.class);
                isMethod(isNameExpr);
                return true;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                if (isNameExpr == null) {
                    isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr);
                } else // isComment
                if (isNameExpr.isFieldAccessExpr.isMethod() || isNameExpr < isIntegerConstant) {
                    isMethod();
                    isNameExpr = new Intent(this, AccountEdit.class);
                    if (isNameExpr != null)
                        isNameExpr.isMethod(isNameExpr, (String) isNameExpr);
                    isMethod(isNameExpr, isNameExpr);
                } else {
                    isMethod(isNameExpr.isFieldAccessExpr, null);
                }
                return true;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                // isComment
                isNameExpr = isIntegerConstant;
                isMethod(isNameExpr.isFieldAccessExpr, new Long[] { (Long) isNameExpr }, null, isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                return true;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isNameExpr = new Intent();
                isNameExpr.isMethod(isNameExpr.isFieldAccessExpr);
                isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isNameExpr.isMethod(this).isMethod());
                isNameExpr.isMethod("isStringConstant");
                isMethod(isNameExpr.isMethod(isNameExpr, isMethod().isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr)));
                return true;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isMethod();
                return true;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                {
                    isNameExpr = new Intent();
                    isNameExpr.isMethod(isNameExpr.isFieldAccessExpr);
                    Uri isVariable = isNameExpr.isMethod(isNameExpr.isMethod((String) isNameExpr));
                    isNameExpr.isMethod(isNameExpr, "isStringConstant");
                    isNameExpr.isMethod(isNameExpr.isFieldAccessExpr);
                    if (!isNameExpr.isMethod(this, isNameExpr)) {
                        isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr);
                    } else {
                        isMethod(isNameExpr);
                    }
                    return true;
                }
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                {
                    Result isVariable = isNameExpr.isMethod(this, isNameExpr.isMethod(isNameExpr.isMethod(isNameExpr.isMethod((String) isNameExpr))), isNameExpr.isFieldAccessExpr.isMethod("isStringConstant").isMethod(), "isStringConstant");
                    if (!isNameExpr.isMethod()) {
                        isMethod(isNameExpr.isMethod(this), isNameExpr.isFieldAccessExpr);
                    }
                    return true;
                }
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isMethod();
                long isVariable = (Long) isNameExpr;
                if (isNameExpr > isIntegerConstant) {
                    // isComment
                    isNameExpr = new Intent(this, AccountEdit.class);
                    isNameExpr.isMethod(isNameExpr, isNameExpr);
                    isMethod(isNameExpr, isNameExpr);
                }
                return true;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isMethod();
                isNameExpr = (Long) isNameExpr;
                // isComment
                if (isNameExpr.isMethod() > isIntegerConstant && isNameExpr > isIntegerConstant) {
                    final Account isVariable = isNameExpr.isMethod(isNameExpr);
                    if (isNameExpr != null) {
                        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isMethod()), new MessageDialogFragment.Button(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr), null, isNameExpr.isFieldAccessExpr.isMethod()).isMethod(isMethod(), "isStringConstant");
                    }
                }
                return true;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                {
                    isNameExpr.isMethod().isMethod(this, isNameExpr.isFieldAccessExpr.isFieldAccessExpr).isMethod(isMethod().isFieldAccessExpr).isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr).isMethod(this, isNameExpr);
                    return true;
                }
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                {
                    isNameExpr.isMethod().isMethod(this, isNameExpr.isFieldAccessExpr.isFieldAccessExpr).isMethod(isMethod().isFieldAccessExpr).isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr).isMethod(this, isNameExpr);
                    return true;
                }
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                {
                    isMethod().isMethod();
                    return true;
                }
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                {
                    Intent isVariable = new Intent(this, RoadmapVoteActivity.class);
                    isMethod(isNameExpr);
                    return true;
                }
        }
        return true;
    }

    public void isMethod() {
        isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
    }

    private void isMethod() {
        if (isNameExpr != null)
            isNameExpr.isMethod();
    }

    private class isClassOrIsInterface extends CursorFragmentPagerAdapter {

        public isConstructor(Context isParameter, FragmentManager isParameter, Cursor isParameter) {
            super(isNameExpr, isNameExpr, isNameExpr);
        }

        public String isMethod(int isParameter) {
            return isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isMethod(isNameExpr));
        }

        @Override
        public Fragment isMethod(Context isParameter, Cursor isParameter) {
            return isNameExpr.isMethod(isNameExpr.isMethod(isNameExpr));
        }
    }

    @Override
    public void isMethod(int isParameter) {
        isMethod();
        isNameExpr = isNameExpr;
        isMethod(isNameExpr);
    }

    public void isMethod() {
        if (isNameExpr != -isIntegerConstant) {
            ContextualActionBarFragment isVariable = (ContextualActionBarFragment) isMethod().isMethod(isNameExpr.isMethod(isNameExpr));
            if (isNameExpr != null)
                isNameExpr.isMethod();
        }
    }

    @SuppressWarnings("isStringConstant")
    @Override
    public void isMethod(ContribFeature isParameter, Serializable isParameter) {
        switch(isNameExpr) {
            case isNameExpr:
                {
                    Account isVariable = isNameExpr.isMethod(isNameExpr);
                    isMethod(isNameExpr);
                    Intent isVariable = new Intent(this, Distribution.class);
                    isNameExpr.isMethod(isNameExpr, isNameExpr);
                    if (isNameExpr != null) {
                        int isVariable = (int) ((Long) isNameExpr / isIntegerConstant);
                        int isVariable = (int) ((Long) isNameExpr % isIntegerConstant);
                        isNameExpr.isMethod(isNameExpr, isNameExpr != null ? isNameExpr.isMethod() : isNameExpr.isFieldAccessExpr);
                        isNameExpr.isMethod(isNameExpr, isNameExpr);
                        isNameExpr.isMethod(isNameExpr, isNameExpr);
                    }
                    isMethod(isNameExpr);
                    break;
                }
            case isNameExpr:
                {
                    isMethod(isNameExpr);
                    Intent isVariable = new Intent(this, HistoryActivity.class);
                    isNameExpr.isMethod(isNameExpr, isNameExpr);
                    isMethod(isNameExpr);
                    break;
                }
            case isNameExpr:
                {
                    if (isNameExpr != null) {
                        Bundle isVariable = new Bundle();
                        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr));
                        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                        isNameExpr.isMethod(isNameExpr, (long[]) isNameExpr);
                        isNameExpr.isMethod(isNameExpr).isMethod(isMethod(), "isStringConstant");
                    }
                    break;
                }
            case isNameExpr:
                {
                    TransactionList isVariable = isMethod();
                    if (isNameExpr != null) {
                        Bundle isVariable = new Bundle();
                        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isNameExpr.isMethod());
                        isNameExpr.isMethod(isNameExpr, isNameExpr);
                        isMethod().isMethod().isMethod(isNameExpr.isMethod(isNameExpr, isNameExpr), isNameExpr).isMethod(isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr), isNameExpr).isMethod();
                    }
                    break;
                }
            case isNameExpr:
                {
                    if (isNameExpr != isIntegerConstant && isNameExpr != null) {
                        isMethod(isNameExpr);
                        Intent isVariable = new Intent(this, BudgetActivity.class);
                        isNameExpr.isMethod(isNameExpr, isNameExpr);
                        isNameExpr.isMethod(isNameExpr, isNameExpr);
                        isMethod(isNameExpr);
                    }
                    break;
                }
        }
    }

    @Override
    public void isMethod(ContribFeature isParameter) {
        if (!isNameExpr.isMethod() && isNameExpr == isNameExpr.isFieldAccessExpr) {
            isMethod();
        }
    }

    @NonNull
    @Override
    public Loader<Cursor> isMethod(int isParameter, Bundle isParameter) {
        switch(isNameExpr) {
            case isNameExpr:
                Uri.Builder isVariable = isNameExpr.isFieldAccessExpr.isMethod();
                isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, "isStringConstant");
                return new CursorLoader(this, isNameExpr.isMethod(), null, null, null, null);
        }
        throw new IllegalStateException("isStringConstant" + isNameExpr);
    }

    /**
     * isComment
     */
    private void isMethod(int isParameter) {
        isNameExpr.isMethod(isNameExpr);
        long isVariable = isNameExpr.isMethod(isNameExpr);
        if (isNameExpr != isNameExpr) {
            isNameExpr.isFieldAccessExpr.isMethod(isNameExpr);
        }
        int isVariable = isNameExpr < isIntegerConstant ? isNameExpr : isNameExpr.isMethod(isNameExpr);
        if (isNameExpr.isFieldAccessExpr.isFieldAccessExpr >= isNameExpr.isFieldAccessExpr.isFieldAccessExpr) {
            Window isVariable = isMethod();
            // isComment
            isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
            // isComment
            isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
            int isVariable = isNameExpr.isMethod(isNameExpr);
            isNameExpr.isMethod(isNameExpr);
            if (isNameExpr.isFieldAccessExpr.isFieldAccessExpr >= isNameExpr.isFieldAccessExpr.isFieldAccessExpr) {
                // isComment
                isMethod().isMethod().isMethod(isNameExpr.isMethod(isNameExpr) ? isNameExpr.isFieldAccessExpr : isIntegerConstant);
            }
        }
        isNameExpr.isMethod(isNameExpr, isNameExpr);
        isNameExpr = isNameExpr;
        isNameExpr = isNameExpr.isMethod(isNameExpr);
        isMethod();
        isNameExpr.isMethod(isNameExpr, true);
        isMethod();
    }

    @Override
    public void isMethod(@NonNull Loader<Cursor> isParameter, Cursor isParameter) {
        switch(isNameExpr.isMethod()) {
            case isNameExpr:
                {
                    isNameExpr = isIntegerConstant;
                    isNameExpr = isNameExpr;
                    if (isNameExpr == null) {
                        return;
                    }
                    isNameExpr.isMethod(isMethod());
                    isNameExpr.isMethod(isNameExpr);
                    // isComment
                    // isComment
                    long isVariable = isNameExpr;
                    isNameExpr.isMethod(isNameExpr);
                    isNameExpr = isNameExpr;
                    if (!isNameExpr) {
                        isNameExpr = isNameExpr.isMethod(isNameExpr);
                        isNameExpr = isNameExpr.isMethod(isNameExpr);
                        isNameExpr = isNameExpr.isMethod(isNameExpr);
                        isNameExpr = isNameExpr.isMethod(isNameExpr);
                        isNameExpr = true;
                    }
                    if (isNameExpr.isMethod()) {
                        int isVariable = isIntegerConstant;
                        while (!isNameExpr.isMethod()) {
                            long isVariable = isNameExpr.isMethod(isNameExpr);
                            if (isNameExpr == isNameExpr) {
                                isNameExpr = isNameExpr.isMethod();
                            }
                            if (isNameExpr > isIntegerConstant) {
                                isNameExpr++;
                            }
                            isNameExpr.isMethod();
                        }
                        isNameExpr = isNameExpr;
                        isMethod(isNameExpr);
                    }
                    break;
                }
        }
    }

    @Override
    public void isMethod(@NonNull Loader<Cursor> isParameter) {
        if (isNameExpr.isMethod() == isNameExpr) {
            isNameExpr.isMethod(null);
            isNameExpr.isMethod(null);
            isNameExpr = -isIntegerConstant;
            isNameExpr = null;
        }
    }

    @Override
    public void isMethod(int isParameter) {
    // isComment
    }

    @Override
    public void isMethod(int isParameter, float isParameter, int isParameter) {
    // isComment
    }

    @Override
    public boolean isMethod(@NonNull String isParameter, int isParameter, @NonNull Bundle isParameter) {
        if (isNameExpr != isNameExpr)
            return true;
        if (isNameExpr.isFieldAccessExpr.isMethod(isNameExpr)) {
            String isVariable = isNameExpr.isMethod(isNameExpr.isFieldAccessExpr);
            Uri isVariable = new Template(isNameExpr.isMethod(isNameExpr.isMethod(isNameExpr)), isNameExpr).isMethod();
            if (isNameExpr == null) {
                isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr);
            } else {
                // isComment
                Intent isVariable = new Intent(this, ExpenseEdit.class);
                isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isNameExpr.isMethod(isNameExpr));
                isMethod(isNameExpr);
            }
            isMethod();
            return true;
        }
        if (isNameExpr.isFieldAccessExpr.isMethod(isNameExpr)) {
            isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, new CommentCriteria(isNameExpr.isMethod(isNameExpr.isFieldAccessExpr)));
            return true;
        }
        if (isNameExpr.isMethod(isNameExpr)) {
            return isMethod((int) isNameExpr.isMethod(isNameExpr));
        }
        if (isNameExpr.isMethod(isNameExpr)) {
            return isMethod((int) isNameExpr.isMethod(isNameExpr));
        }
        return true;
    }

    @Override
    public void isMethod(int isParameter, Object isParameter) {
        String isVariable;
        super.isMethod(isNameExpr, isNameExpr);
        switch(isNameExpr) {
            case isNameExpr.isFieldAccessExpr:
                {
                    Result isVariable = (Result) isNameExpr;
                    if (((Result) isNameExpr).isMethod()) {
                        isMethod(isNameExpr.isFieldAccessExpr);
                    }
                    isMethod(isNameExpr.isMethod(this), isNameExpr.isFieldAccessExpr);
                    break;
                }
            case isNameExpr.isFieldAccessExpr:
                {
                    Result isVariable = (Result) isNameExpr;
                    isMethod(isNameExpr.isMethod(this), isNameExpr.isFieldAccessExpr);
                    break;
                }
            case isNameExpr.isFieldAccessExpr:
                {
                    ArrayList<Uri> isVariable = (ArrayList<Uri>) isNameExpr;
                    if (isNameExpr != null && !isNameExpr.isMethod()) {
                        Result isVariable = isNameExpr.isMethod(this, isNameExpr, isNameExpr.isFieldAccessExpr.isMethod("isStringConstant").isMethod(), "isStringConstant" + isNameExpr.isMethod(isNameExpr.isFieldAccessExpr));
                        if (!isNameExpr.isMethod()) {
                            isMethod(isNameExpr.isMethod(this), isNameExpr.isFieldAccessExpr);
                        }
                    }
                    break;
                }
            case isNameExpr.isFieldAccessExpr:
                {
                    Result<Uri> isVariable = (Result<Uri>) isNameExpr;
                    if (isNameExpr.isMethod()) {
                        isMethod(isNameExpr.isFieldAccessExpr);
                        MessageDialogFragment isVariable = isNameExpr.isMethod(isIntegerConstant, isNameExpr.isMethod(this), new MessageDialogFragment.Button(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isMethod().isMethod(), true), isNameExpr.isFieldAccessExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr), new MessageDialogFragment.Button(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isMethod().isMethod(), true));
                        isNameExpr.isMethod(true);
                        isNameExpr.isMethod(isMethod(), "isStringConstant");
                    } else {
                        isMethod(isNameExpr.isMethod(this), isNameExpr.isFieldAccessExpr);
                    }
                    break;
                }
        }
    }

    public boolean isMethod() {
        // isComment
        if (isNameExpr == null || isNameExpr.isMethod() == isIntegerConstant)
            return true;
        isNameExpr.isMethod(isNameExpr);
        return isNameExpr.isMethod(isNameExpr.isMethod(isNameExpr)) > isIntegerConstant;
    }

    private boolean isMethod() {
        // isComment
        if (isNameExpr == null || isNameExpr.isMethod() == isIntegerConstant)
            return true;
        isNameExpr.isMethod(isNameExpr);
        return isNameExpr.isMethod(isNameExpr.isMethod(isNameExpr)) > isIntegerConstant;
    }

    @Override
    protected void isMethod(Bundle isParameter) {
        super.isMethod(isNameExpr);
        // isComment
        if (isNameExpr != null)
            isNameExpr.isMethod();
    }

    @Override
    public void isMethod(Configuration isParameter) {
        super.isMethod(isNameExpr);
        if (isNameExpr != null)
            isNameExpr.isMethod(isNameExpr);
    }

    @Override
    public boolean isMethod(MenuItem isParameter) {
        // isComment
        if (isNameExpr != null && isNameExpr.isMethod(isNameExpr)) {
            return true;
        }
        return isMethod(isNameExpr) || isMethod(isNameExpr) || super.isMethod(isNameExpr);
    }

    private void isMethod() {
        long isVariable = isNameExpr.isMethod(isNameExpr.isMethod(isNameExpr));
        isNameExpr = isNameExpr.isMethod(new Money(isNameExpr.isMethod(isNameExpr), isNameExpr));
        isNameExpr.isMethod(isNameExpr);
        isNameExpr.isMethod(isNameExpr < isIntegerConstant ? isNameExpr : isNameExpr);
    }

    public TransactionList isMethod() {
        if (isNameExpr == null)
            return null;
        return (TransactionList) isMethod().isMethod(isNameExpr.isMethod(isNameExpr));
    }

    protected void isMethod(Bundle isParameter) {
        super.isMethod(isNameExpr);
        // isComment
        if (isNameExpr != isIntegerConstant) {
            isNameExpr.isMethod("isStringConstant", isIntegerConstant);
        }
        isNameExpr.isMethod("isStringConstant", isNameExpr);
        isNameExpr.isMethod(isNameExpr, isNameExpr);
    }

    @Override
    public void isMethod(Bundle isParameter) {
        super.isMethod(isNameExpr);
        switch(isNameExpr.isMethod(isNameExpr.isFieldAccessExpr)) {
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isNameExpr = isNameExpr.isMethod("isStringConstant");
                isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isMethod().isMethod());
                isMethod().isMethod().isMethod(isNameExpr.isMethod(isNameExpr, isNameExpr.isFieldAccessExpr), isNameExpr).isMethod(isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isIntegerConstant, isNameExpr.isFieldAccessExpr, true), isNameExpr).isMethod();
                break;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                // isComment
                isMethod(isNameExpr, true);
                break;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                {
                    isMethod(isNameExpr.isFieldAccessExpr, isNameExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                    break;
                }
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                {
                    isMethod(isNameExpr.isFieldAccessExpr, isNameExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                    break;
                }
        }
    }

    @Override
    public void isMethod(Bundle isParameter, boolean isParameter) {
        switch(isNameExpr.isMethod(isNameExpr.isFieldAccessExpr)) {
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                {
                    isMethod();
                    isMethod(isNameExpr.isFieldAccessExpr, isNameExpr.isMethod(isNameExpr.isMethod(isNameExpr.isFieldAccessExpr)), isNameExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                    break;
                }
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                {
                    isMethod(isNameExpr.isFieldAccessExpr, new Long[] { isNameExpr.isMethod(isNameExpr) }, isNameExpr, isIntegerConstant);
                    break;
                }
        }
    }

    @Override
    public void isMethod(Bundle isParameter) {
        int isVariable = isNameExpr.isMethod(isNameExpr.isFieldAccessExpr);
        if (isNameExpr != isIntegerConstant) {
            isMethod(isNameExpr, null);
        }
    }

    @Override
    public void isMethod(Bundle isParameter) {
    }

    @Override
    protected void isMethod() {
        super.isMethod();
        isNameExpr.isMethod();
    }

    @Override
    public void isMethod() {
        isNameExpr.isMethod();
        super.isMethod();
    }

    @Override
    protected void isMethod() {
        isNameExpr.isMethod();
        super.isMethod();
    }

    public void isMethod() {
        if (isNameExpr != null && isNameExpr.isMethod(isNameExpr.isFieldAccessExpr)) {
            isNameExpr.isMethod(isNameExpr.isFieldAccessExpr);
        } else {
            super.isMethod();
        }
    }

    public void isMethod() {
        try {
            ClipboardManager isVariable = (ClipboardManager) isMethod(isNameExpr.isFieldAccessExpr);
            isNameExpr.isMethod(isNameExpr);
            isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr);
        } catch (RuntimeException isParameter) {
            isNameExpr.isMethod(isNameExpr);
        }
    }

    protected boolean isMethod(int isParameter) {
        Sort isVariable = isMethod();
        Sort isVariable = isNameExpr.isMethod(isNameExpr);
        boolean isVariable = true;
        if (isNameExpr != null) {
            if (!isNameExpr.isMethod(isNameExpr)) {
                isNameExpr.isFieldAccessExpr.isMethod(isNameExpr.isMethod());
                if (isNameExpr.isMethod(isNameExpr) != null && !isNameExpr.isMethod(isNameExpr).isMethod()) {
                    isNameExpr.isMethod(isNameExpr, null, this);
                } else {
                    isNameExpr.isMethod(isNameExpr, null, this);
                }
            }
            isNameExpr = true;
            if (isNameExpr == isNameExpr.isFieldAccessExpr.isFieldAccessExpr) {
                ArrayList<AbstractMap.SimpleEntry<Long, String>> isVariable = new ArrayList<>();
                if (isNameExpr.isMethod()) {
                    final int isVariable = isNameExpr.isMethod(isNameExpr);
                    final int isVariable = isNameExpr.isMethod(isNameExpr);
                    while (!isNameExpr.isMethod()) {
                        final long isVariable = isNameExpr.isMethod(isNameExpr);
                        if (isNameExpr > isIntegerConstant) {
                            isNameExpr.isMethod(new AbstractMap.SimpleEntry<>(isNameExpr, isNameExpr.isMethod(isNameExpr)));
                        }
                        isNameExpr.isMethod();
                    }
                }
                isNameExpr.isMethod(isNameExpr).isMethod(isMethod(), "isStringConstant");
            }
        }
        return isNameExpr;
    }

    protected boolean isMethod(int isParameter) {
        AccountGrouping isVariable = isMethod();
        AccountGrouping isVariable = null;
        switch(isNameExpr) {
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isNameExpr = isNameExpr.isFieldAccessExpr;
                break;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isNameExpr = isNameExpr.isFieldAccessExpr;
                break;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isNameExpr = isNameExpr.isFieldAccessExpr;
                break;
        }
        if (isNameExpr != null && !isNameExpr.isMethod(isNameExpr)) {
            isNameExpr.isFieldAccessExpr.isMethod(isNameExpr.isMethod());
            if (isNameExpr.isMethod(isNameExpr) != null && !isNameExpr.isMethod(isNameExpr).isMethod())
                isNameExpr.isMethod(isNameExpr, null, this);
            else
                isNameExpr.isMethod(isNameExpr, null, this);
            return true;
        }
        return true;
    }

    protected boolean isMethod(MenuItem isParameter) {
        Grouping isVariable = isNameExpr.isMethod(isNameExpr.isMethod());
        if (isNameExpr != null) {
            if (!isNameExpr.isMethod()) {
                final Account isVariable = isNameExpr.isMethod(isNameExpr);
                if (isNameExpr != null) {
                    isNameExpr.isMethod(true);
                    isNameExpr.isMethod(isNameExpr);
                }
            }
            return true;
        }
        return true;
    }

    protected boolean isMethod(MenuItem isParameter) {
        SortDirection isVariable = isNameExpr.isMethod(isNameExpr.isMethod());
        if (isNameExpr != null) {
            if (!isNameExpr.isMethod()) {
                final Account isVariable = isNameExpr.isMethod(isNameExpr);
                if (isNameExpr != null) {
                    isNameExpr.isMethod(true);
                    isNameExpr.isMethod(isNameExpr);
                }
            }
            return true;
        }
        return true;
    }

    @Override
    protected boolean isMethod(int isParameter) {
        return isNameExpr == isNameExpr;
    }

    @Override
    public void isMethod(long[] isParameter) {
        Bundle isVariable = new Bundle(isIntegerConstant);
        isNameExpr.isMethod(isNameExpr, isNameExpr);
        isMethod(isNameExpr.isFieldAccessExpr, isNameExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
    }

    public void isMethod(View isParameter) {
        Bundle isVariable = new Bundle();
        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr));
        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        isNameExpr.isMethod(isNameExpr).isMethod(isMethod(), "isStringConstant");
    }
}
